# docs/api/openapi/plan-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "혼디가개 여행 일정 서비스 API 명세서",
    "description": "Plan Service 전용",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Plan Service"
    }
  ],
...
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "tags": [
          "여행 일정"
        ],
        "summary": "내 여행 일정 목록 조회",
        "description": "커서 기반으로 내 여행 일정 목록을 조회합니다. 일정 아이디 내림차순(최근 생성순)으로 내려갑니다.\n\n**필수 파라미터는 없습니다.** 전부 생략하면 내 일정 전체의 첫 페이지(10개)가 옵니다.\n\n호출 예\n- 첫 페이지: `GET /api/v1/plans`\n- 특정 반려견이 동행한 일정만: `GET /api/v1/plans?petId=1234567890123456789`\n- 다음 페이지: 직전 응답 `items` 마지막의 `planId` 를 `lastPlanId` 로 — `GET /api/v1/plans?lastPlanId=1234567890123456789&size=10`",
        "operationId": "getMyPlans",
        "parameters": [
          {
            "name": "petId",
            "in": "query",
            "description": "[선택] 반려견 아이디. 지정하면 그 반려견이 동행한 일정만 조회합니다 (반려견별 여행 히스토리). 여러 마리 일정은 그중 한 마리로 들어 있어도 조회됩니다. 생략하면 내 일정 전체. Snowflake 라 환경마다 다르고 예시 값은 형식 안내용, 실제 값은 반려견 목록 응답의 petId 를 씁니다",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1234567890123456800
          },
          {
            "name": "lastPlanId",
            "in": "query",
            "description": "[선택] 커서. 첫 페이지는 생략하고, 다음 페이지는 직전 응답 마지막 항목의 planId 를 넣습니다. 예시 값은 형식 안내용",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1234567890123456800
          },
          {
            "name": "size",
            "in": "query",
            "description": "[선택, 기본 10] 조회 개수 (1~50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "maximum": 50,
              "minimum": 1
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSliceResponsePlanSummaryItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "여행 일정"
        ],
        "summary": "여행 일정 생성",
        "description": "여행 일정을 생성합니다. 일정 항목을 함께 보내면 같이 저장하며, 장소 항목은 tour-service에서 존재를 검증합니다. 동행 반려견은 petIds(최대 5마리)로 보내고 첫 번째가 대표 반려견이 됩니다. petIds 가 없으면 petId 를, 둘 다 없으면 대표 반려견을 씁니다 — AI 일정 생성(POST /ai-plans)과 같은 규칙입니다.\n\n**필수: 요청 바디의 areaCode, title, startDate, endDate.** petId·petIds·sigunguCode·budget·items 는 생략 가능하고, 여행 기간(startDate~endDate)은 최대 30일입니다. items 를 보낼 때는 각 항목의 day·itemType·title 이 필수입니다.\n\n호출 예\n- 최소 바디(대표 반려견, 항목 없음): `POST /api/v1/plans` `{\"areaCode\":\"39\",\"title\":\"몽실이와 제주 2박 3일\",\"startDate\":\"2026-09-12\",\"endDate\":\"2026-09-14\"}`\n- 반려견 두 마리 지정: 위 바디에 `\"petIds\":[1234567890123456789,1234567890123456790]` 추가",
        "operationId": "createPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePlanDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/favorites/places/{placeId}": {
      "get": {
        "tags": [
          "장소 즐겨찾기"
        ],
        "summary": "즐겨찾기 여부 확인",
        "description": "장소 한 곳의 즐겨찾기 저장 여부를 확인합니다. 상세 화면의 토글 초기 상태용입니다.\n\n**필수: placeId (경로).**\n\n호출 예\n- `GET /api/v1/favorites/places/212481712381923328`",
        "operationId": "getFavoriteStatus",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "[필수] 장소 아이디. Snowflake 숫자라 환경(dev/prod)마다 다르고 예시 값은 형식 안내용입니다. 실제 값은 tour-service 장소 목록·주변 조회 응답의 placeId 를 그대로 씁니다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 212481712381923330
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseFavoriteStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "장소 즐겨찾기"
        ],
        "summary": "즐겨찾기 저장",
        "description": "장소를 즐겨찾기에 저장합니다. 이미 저장된 장소면 그대로 성공합니다(멱등). 회원당 최대 100곳까지 저장할 수 있습니다.\n\n**필수: placeId (경로).** 요청 바디는 없습니다.\n\n호출 예\n- `POST /api/v1/favorites/places/212481712381923328`",
        "operationId": "addFavorite",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "[필수] 장소 아이디. Snowflake 숫자라 환경(dev/prod)마다 다르고 예시 값은 형식 안내용입니다. 실제 값은 tour-service 장소 목록·주변 조회 응답의 placeId 를 그대로 씁니다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 212481712381923330
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "장소 즐겨찾기"
        ],
        "summary": "즐겨찾기 해제",
        "description": "장소를 즐겨찾기에서 제거합니다. 저장되어 있지 않아도 성공합니다(멱등).\n\n**필수: placeId (경로).**\n\n호출 예\n- `DELETE /api/v1/favorites/places/212481712381923328`",
        "operationId": "removeFavorite",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "[필수] 장소 아이디. Snowflake 숫자라 환경(dev/prod)마다 다르고 예시 값은 형식 안내용입니다. 실제 값은 tour-service 장소 목록·주변 조회 응답의 placeId 를 그대로 씁니다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 212481712381923330
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/plans/{planId}/weather": {
      "get": {
        "tags": [
          "여행 일정"
        ],
        "summary": "일정 날씨 브리핑",
        "description": "일정의 날짜별로 날씨와 반려견 여행 적합도를 같이 보여 줍니다. 그날 첫 장소 항목을 기준으로 판정하며, 반려견 특성은 등록된 프로필을 자동으로 반영합니다. 여러 마리 일정은 아이별로 따로 판정하고 점수가 가장 낮은 아이를 그날의 기준(basisPetId)으로 삼습니다 — 아이별 점수는 petSuitabilities 에 함께 내려갑니다. 예보가 닿는 범위는 오늘부터 오늘+10일(11일)이며, 그 밖의 날짜는 score 가 null 입니다 — 점수가 낮은 것이 아니라 판단 근거가 없는 것입니다. 판정을 못 낸 날은 unavailableReasonCode 와 unavailableReason(문장)이 짝으로 옵니다. 코드는 넷으로 갈립니다 — PAST_DATE(지난 날짜, 다시 물어도 생기지 않습니다) · NO_PLACE_ITEM(그날 장소 항목 없음) · BEYOND_FORECAST_RANGE(예보 범위 밖) · LOOKUP_FAILED(조회 실패, 넷 중 이것만 일시적 장애입니다). 재시도 안내는 이 코드로 판단하고 문장을 파싱하지 않습니다. 비 예보가 있고 그날 장소가 실내가 아니면 indoorAlternatives 에 실내 대안을 함께 내려 줍니다. 일정을 실제로 바꾸는 것은 이 API 가 아니라 일자별 항목 교체 API 로 명시적으로 합니다.\n\n**필수: planId (경로).**\n\n호출 예\n- `GET /api/v1/plans/1234567890123456789/weather`",
        "operationId": "getPlanWeather",
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "description": "[필수] 일정 아이디. Snowflake 숫자라 환경(dev/prod)마다 다르고 예시 값은 형식 안내용입니다. 실제 값은 일정 목록 응답의 planId 를 그대로 씁니다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1234567890123456800
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePlanWeatherResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/plans/{planId}/walk-safety": {
      "get": {
        "tags": [
          "일정 산책 위험도"
        ],
        "summary": "일정 항목 산책 위험도",
        "description": "일정 항목마다 **그 항목의 시각 기준으로** 산책 위험도를 판정합니다. 일자 날씨 브리핑(`/weather`)이 \"둘째 날 괜찮아?\" 에 답한다면 이 API 는 \"두 시에 그 해수욕장 걸어도 돼?\" 에 답합니다 — 산책 위험도는 시각에 따라 갈려 일자로 접을 수 없습니다. 판정 규칙(노면온도 추정·체감온도·등급 임계)은 장소 산책 위험도 API 와 **같은 것**을 씁니다. 판정 기준 반려견(basisPetId)은 **그날 날씨 판정의 기준과 같습니다** — 두 화면이 같은 날을 다른 아이 기준으로 말하지 않게 합니다.\n\n판정을 못 낸 항목도 줄은 그대로 내려가고 unavailableReasonCode 와 unavailableReason(문장)이 짝으로 옵니다. 코드는 다섯으로 갈립니다 — PAST_DATE(지난 날짜, 다시 물어도 생기지 않습니다) · NOT_PLACE_TARGET(좌표를 아는 장소 항목이 아님, 예: 올레 코스·이동) · NO_START_TIME(항목에 시각이 없음 — **없는 시각을 정오 등으로 지어내지 않습니다**) · BEYOND_FORECAST_RANGE(시각별 예보 범위 밖, 기다리면 풀립니다) · LOOKUP_FAILED(조회 실패, 다섯 중 이것만 일시적 장애입니다). 재시도 안내는 이 코드로 판단하고 문장을 파싱하지 않습니다.\n\n**시각별 예보가 닿는 범위는 `오늘 ~ 오늘+4`(5일)로, 일자 날씨 브리핑의 11일보다 짧습니다.** 산책 위험도는 노면온도를 시각·일사로 추정해야 해서 시각별 데이터가 있는 단기예보만 쓰고, 중기예보(오늘+4~오늘+10)에는 오전/오후뿐이라 오후 두 시 아스팔트를 계산할 수 없기 때문입니다. 그 밖의 날짜는 원격 호출 없이 BEYOND_FORECAST_RANGE 로 내려갑니다 — 같은 일정이라도 `/weather` 에는 판정이 있는데 여기는 비어 있는 날이 **정상**입니다.\n\n근거 목록·시간대별 판정·특보 상세는 이 응답에 없습니다 — 항목마다 실어 나르면 응답이 항목 수만큼 부풀기 때문입니다. 필요하면 장소 산책 위험도 API(`GET /api/v1/places/{placeId}/walk-safety`)를 직접 부르세요. 판정을 못 낸 줄에도 placeId 는 남습니다(NOT_PLACE_TARGET 제외) — 그래야 그 줄에서 직접 부를 수 있습니다.\n\n**필수: planId (경로).**\n\n호출 예\n- `GET /api/v1/plans/1234567890123456789/walk-safety`",
        "operationId": "getWalkSafety",
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "description": "[필수] 일정 아이디. Snowflake 숫자라 환경(dev/prod)마다 다르고 예시 값은 형식 안내용입니다. 실제 값은 일정 목록 응답의 planId 를 그대로 씁니다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1234567890123456800
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePlanWalkSafetyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/favorites/places": {
      "get": {
        "tags": [
          "장소 즐겨찾기"
        ],
        "summary": "내 즐겨찾기 목록",
        "description": "저장한 장소를 최근 저장순으로 조회합니다. 장소 요약은 tour-service 에서 붙이며, 조회에 실패해도 placeId 목록은 내려갑니다.\n\n**필수 파라미터는 없습니다.** 페이징 없이 저장한 장소 전체(회원당 최대 100곳)가 한 번에 옵니다.\n\n호출 예\n- `GET /api/v1/favorites/places`",
        "operationId": "getMyFavorites",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseFavoritePlacesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataHeader": {
        "type": "object",
        "description": "공통 응답 헤더",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "성공 여부"
          },
          "resultCode": {
            "type": "string",
            "description": "대표 오류 코드. 성공이면 null"
          },
          "resultMessage": {
            "type": "string",
            "description": "사용자에게 보여줄 대표 오류 메시지. 성공이면 null. 오류 종류와 무관하게 항상 문자열이다"
          },
          "fieldErrors": {
            "type": "array",
            "description": "필드 단위 검증 오류 목록. 검증 실패가 아니면 null",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        }
      },
      "PlanDetailResponse": {
        "type": "object",
        "description": "여행 일정 상세 응답 DTO",
        "properties": {
          "planId": {
            "type": "string",
            "description": "일정 아이디",
            "example": 1234567890123456800
          },
          "petId": {
            "type": "string",
            "description": "대표 반려견 아이디 — petIds 의 첫 번째와 같다",
            "example": 1234567890123456800
          },
          "petIds": {
            "type": "array",
            "description": "동행 반려견 아이디 목록 (한 마리 일정이면 원소 하나)",
            "example": [
              "1234567890123456789"
            ],
            "items": {
              "type": "string"
            }
          },
          "areaCode": {
            "type": "string",
            "description": "지역 코드",
            "example": 39
          },
          "sigunguCode": {
            "type": "string",
            "description": "시군구 코드",
            "example": 4
          },
          "title": {
            "type": "string",
            "description": "일정 제목",
            "example": "몽실이와 제주 2박 3일"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "여행 시작일",
            "example": "2026-09-12"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "여행 종료일",
            "example": "2026-09-14"
          },
          "budget": {
            "type": "integer",
            "format": "int32",
            "description": "예산 (원)",
            "example": 400000
          },
          "status": {
            "$ref": "#/components/schemas/CodeNameDescriptionMetadata",
            "description": "일정 상태",
            "example": {
              "code": "DRAFT",
              "name": "초안",
              "description": "AI 또는 사용자가 작성 중인 일정입니다."
            }
          },
          "totalDays": {
            "type": "integer",
            "format": "int32",
            "description": "총 여행 일수",
            "example": 3
          },
          "items": {
            "type": "array",
            "description": "일정 항목 목록 (일차·순서 오름차순)",
            "items": {
              "$ref": "#/components/schemas/PlanItemDetailItem"
            }
          }
        }
      },
...
      "ResponsePlanDetailResponse": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/PlanDetailResponse",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
...
      "ResponseVoid": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "type": "object",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
...
      "PlanItemRequest": {
        "type": "object",
        "description": "여행 일정 항목 요청 DTO",
        "properties": {
          "day": {
            "type": "integer",
            "format": "int32",
            "description": "일차 (1부터). 일정 생성 요청에서는 필수, 일자별 일괄 교체 요청에서는 경로의 day 로 덮어쓰므로 생략 가능",
            "example": 1,
            "minimum": 1
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "description": "생략 가능. 같은 일차 내 표시 순서(0부터). 생략하면 0이고, 같은 일차 안에서 겹치면 실패합니다.",
            "example": 0
          },
          "itemType": {
            "type": "string",
            "description": "항목 유형. PLACE 장소 · MEAL 식사 · LODGING 숙박 · WALK 산책 · MOVE 이동",
            "enum": [
              "PLACE",
              "MEAL",
              "LODGING",
              "WALK",
              "MOVE"
            ],
            "example": "PLACE"
          },
          "targetId": {
            "type": "integer",
            "format": "int64",
            "description": "생략 가능. 대상 아이디 — PLACE/MEAL/LODGING 은 place.id(tour-service 에서 존재 검증), WALK 는 walk_course.id, MOVE 는 비워 둡니다. 생략하면 대상 없는 항목으로 저장됩니다. Snowflake 라 예시 값은 형식 안내용",
            "example": 212481712381923330
          },
          "title": {
            "type": "string",
            "description": "항목 이름 (100자 이하)",
            "example": "천지연폭포",
            "maxLength": 100,
            "minLength": 0
          },
          "memo": {
            "type": "string",
            "description": "생략 가능. 메모(500자 이하). 생략하면 비워 둡니다.",
            "example": "그늘이 많아 더위에 약한 아이도 괜찮음",
            "maxLength": 500,
            "minLength": 0
          },
          "startTime": {
            "type": "string",
            "description": "생략 가능. 시작 시각 (HH:mm:ss). 생략하면 비워 둡니다.",
            "example": "10:30:00"
          }
        },
        "required": [
          "itemType",
          "title"
        ]
      },
      "PlanCreateRequest": {
        "type": "object",
        "description": "여행 일정 생성 요청 DTO",
        "properties": {
          "petId": {
            "type": "integer",
            "format": "int64",
            "description": "생략 가능. 대상 반려견 아이디 — petIds 가 있으면 무시되고, 둘 다 생략하면 대표 반려견이 쓰입니다.",
            "example": 1234567890123456800
          },
          "petIds": {
            "type": "array",
            "description": "생략 가능. 동행 반려견 아이디 목록(최대 5마리) — 첫 번째가 대표 반려견이 되고, 생략하면 petId 또는 대표 반려견을 씁니다.",
            "example": [
              1,
              2
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "maxItems": 5,
            "minItems": 0
          },
          "areaCode": {
            "type": "string",
            "description": "지역 코드 (제주=39)",
            "example": 39,
            "minLength": 1
          },
          "sigunguCode": {
            "type": "string",
            "description": "생략 가능. 시군구 코드(TourAPI sigunguCode). 생략하면 비워 둡니다.",
            "example": 4
          },
          "title": {
            "type": "string",
            "description": "일정 제목 (60자 이하)",
            "example": "몽실이와 제주 2박 3일",
            "maxLength": 60,
            "minLength": 0
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "여행 시작일 (yyyy-MM-dd)",
            "example": "2026-09-12"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "여행 종료일 (yyyy-MM-dd). 시작일과 같거나 이후여야 하고 기간은 최대 30일입니다.",
            "example": "2026-09-14"
          },
          "budget": {
            "type": "integer",
            "format": "int32",
            "description": "생략 가능. 예산(원, 0 이상). 생략하면 비워 둡니다.",
            "example": 400000
          },
          "items": {
            "type": "array",
            "description": "생략 가능. 일정 항목 목록. 생략하면 항목 없는 일정으로 생성되고, 보낼 때는 각 항목의 day 가 필수입니다.",
            "items": {
              "$ref": "#/components/schemas/PlanItemRequest"
            }
          }
        },
        "required": [
          "areaCode",
          "endDate",
          "startDate",
          "title"
        ]
      },
...
      "ResponseSliceResponsePlanSummaryItem": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/SliceResponsePlanSummaryItem",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
      "SliceResponsePlanSummaryItem": {
        "type": "object",
        "properties": {
          "contents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanSummaryItem"
            }
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
...
      "PlanWeatherResponse": {
        "type": "object",
        "description": "일정 날씨 브리핑 응답 DTO",
        "properties": {
          "planId": {
            "type": "string",
            "description": "일정 아이디",
            "example": 1234567890123456800
          },
          "planTitle": {
            "type": "string",
            "description": "일정 제목",
            "example": "몽실이와 제주 2박 3일"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "여행 시작일",
            "example": "2026-09-12"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "여행 종료일",
            "example": "2026-09-14"
          },
          "petIds": {
            "type": "array",
            "description": "판정에 들어간 동행 반려견 아이디 목록. 일자별 basisPetId 는 이 안의 하나다",
            "example": [
              "1234567890123456789"
            ],
            "items": {
              "type": "string"
            }
          },
          "petConditionApplied": {
            "type": "boolean",
            "description": "반려견 특성이 한 마리라도 판정에 반영됐는지. false 면 특성 조회에 실패해 일반 조건으로 판정한 결과다",
            "example": true
          },
          "days": {
            "type": "array",
            "description": "일자별 브리핑. 일정 일수만큼 항상 채워진다",
            "items": {
              "$ref": "#/components/schemas/PlanDayWeatherItem"
            }
          }
        }
      },
      "ResponsePlanWeatherResponse": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/PlanWeatherResponse",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
...
      "PlanWalkSafetyResponse": {
        "type": "object",
        "description": "일정 항목 산책 위험도 응답 DTO",
        "properties": {
          "planId": {
            "type": "string",
            "description": "일정 아이디",
            "example": 1234567890123456800
          },
          "planTitle": {
            "type": "string",
            "description": "일정 제목",
            "example": "몽실이와 제주 2박 3일"
          },
          "petIds": {
            "type": "array",
            "description": "판정에 들어간 동행 반려견 아이디 목록. 항목별 basisPetId 는 이 안의 하나다",
            "example": [
              "1234567890123456789"
            ],
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "description": "일차·순서대로 정렬된 항목별 판정. 못 낸 항목도 사유와 함께 그대로 들어간다",
            "items": {
              "$ref": "#/components/schemas/PlanItemWalkSafetyItem"
            }
          }
        }
      },
      "ResponsePlanWalkSafetyResponse": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/PlanWalkSafetyResponse",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
...
      "ResponseFavoritePlacesResponse": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/FavoritePlacesResponse",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
      "FavoriteStatusResponse": {
        "type": "object",
        "description": "즐겨찾기 여부 응답 DTO",
        "properties": {
          "placeId": {
            "type": "string",
            "description": "장소 아이디",
            "example": 212481712381923330
          },
          "favorited": {
            "type": "boolean",
            "description": "즐겨찾기 저장 여부",
            "example": true
          }
        }
      },
      "ResponseFavoriteStatusResponse": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/FavoriteStatusResponse",
            "description": "응답 본문. 실패면 null"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}